requirements.txt
# =========================
# 语稿智能整理系统依赖包
# =========================

# ------ LLM SDK ------
openai>=1.23.0
anthropic>=0.18.0
qianfan>=0.3.0
dashscope>=1.17.0
zhipuai>=2.0.0

# ------ 本地 LLM （可选）------
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
tokenizers>=0.15.0

# ------ 文档处理 ------
python-docx>=1.1.0
markdown>=3.5.0
docxcompose>=1.4.0
reportlab>=4.0.0  # PDF 生成

# ------ 工具库 ------
pyyaml>=6.0
tqdm>=4.66.0
pydantic>=2.0,<3.0
python-dotenv>=1.0.0
flask>=3.0.0  # Web 框架
flask-cors>=4.0.0  # CORS 支持
requests>=2.31.0  # HTTP 请求（Ollama 支持）

# ------ 文本处理 ------
jieba>=0.42.1
regex>=2023.10.3
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ------ 通用依赖 ------
numpy>=1.26.0
pandas>=2.0.0
